Move evaluation methods into a separate sub-module
