+++ lib/ldclient-rb.rb
@@ -9,3 +9,4 @@ require "ldclient-rb/polling"
 require "ldclient-rb/events"
 require "ldclient-rb/feature_store"
 require "ldclient-rb/requestor"
+require "ldclient-rb/evaluation"
